apps/news/api/v1/views.py
from django_filters.rest_framework import DjangoFilterBackend
from rest_framework import status
from rest_framework.generics import GenericAPIView, ListAPIView
from rest_framework.request import Request
from rest_framework.response import Response

from apps.core.pagination import CustomPagination
from apps.core.utils import GetObjectMixin
from apps.news.api.v1.serializers import ArticleSerializer, CategorySerializer
from apps.news.filters import ArticlesFilter
from apps.news.models import Article, Category


class ArticlesAPIView(GetObjectMixin, GenericAPIView):
    queryset = Article.objects.all()
    serializer_class = ArticleSerializer
    filter_backends = [DjangoFilterBackend]
    filterset_class = ArticlesFilter
    pagination_class = CustomPagination

    def get(self, request: Request) -> Response:

        page = self.paginate_queryset(self.get_queryset())
        if page is not None:
            serializer = self.get_serializer(page, many=True)
            return self.get_paginated_response(serializer.data)

        serializer = self.get_serializer(self.get_queryset(), many=True)
        return Response(serializer.data, status=status.HTTP_200_OK)


class CategoriesAPIView(GetObjectMixin, ListAPIView):
    queryset = Category.objects.all()
    serializer_class = CategorySerializer
    pagination_class = CustomPagination

    def get(self, request: Request) -> Response:

        page = self.paginate_queryset(self.get_queryset())
        if page is not None:
            serializer = self.get_serializer(page, many=True)
            return self.get_paginated_response(serializer.data)

        serializer = self.get_serializer(self.get_queryset(), many=True)
        return Response(serializer.data, status=status.HTTP_200_OK)
