pyproject.toml
[tool.poetry]
name = "cocoon_news_project"
description = "cocoon_news_project"
version = "0.1.0"
readme = "README.md"
authors = ["Emmanuel Okoye <chuckzokoye@gmail.com>"]


[tool.poetry.dependencies]
python = "3.10.0"

django = "^4.1.6"
djangorestframework = "^3.14.0"
django-ckeditor = "^6.5.1"
whitenoise = "^6.3.0"
django-filter = "^22.1"
drf-spectacular = "^0.25.1"


[tool.poetry.dev-dependencies]
factory_boy = "3.2.1"
pylint = "2.13.7"
pytest = "7.1.2"
pytest-cov = "3.0.0"
pytest-django = "4.5.2"
pytest-xdist = "2.5.0"
pytest-socket = "0.5.1"
ptpython = "3.0.20"
mock = "4.0.3"
black = "22.3.0"
ipython = "8.2.0"
django-webtest = "1.9.10"
mypy = "0.981"
django-stubs = "1.12.0"
djangorestframework-stubs = "1.7.0"
freezegun = "1.2.1"
flake8 = "3.8.4"
isort = "5.10.1"
semgrep = "1.1.0"
types-python-dateutil = "2.8.12"
types-six = "1.16.15"
types-requests = "2.27.20"
debugpy = "1.6.0"

[tool.djlint]
profile="django"